RSpace: Introduce `ContResult` data type representing matched continuation.
`Result(value: TaggedContinuation, persist: Boolean)` was not enough to properly refund for clearing the tuplespace. When charging for storing continuation we charge for cost of storing: channels, patterns and body of continuation. `TaggedContinuation` didn't contain first two (channels, patterns). In order to refund equally to what we charged in the first place we need to return those data well.